rename all classes
