+++ education/src/main/java/ru/breev/se/homework2/ArrayExamle.java
